Refactor: Consolidate utility functions into libs/util package
